% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sumZscore.R
\name{calc_sumZscore}
\alias{calc_sumZscore}
\title{Generate Sum Z Score}
\usage{
calc_sumZscore(df, sig, topn = length(sig), botn = NA)
}
\arguments{
\item{df}{numeric dataframe (e.g gene expression df); rownames should contain name of genes/features in signature}

\item{sig}{vector/list of genes (if applicable, list should be ordered so enriched genes are at the top & de-enriched genes are at the bottom)}

\item{topn}{numeric; genes from top of signature to consider}

\item{botn}{numeric; genes from bottom of signature to consider}
}
\value{
scores samples using signature genes
}
\description{
Slightly modified from Favour's function for OV project
}
