% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cbind.fill}
\alias{cbind.fill}
\title{Combine data.frames by column, filling in missing rows.}
\usage{
\method{cbind}{fill}(...)
}
\arguments{
\item{...}{input data frames to column bind together}
}
\value{
a single data frame
}
\description{
\code{cbinds} a list of dataframes filling missing rows with NA.
}
