% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{check_normal}
\alias{check_normal}
\title{Shapiro-Wilk test of normality}
\usage{
check_normal(values)
}
\arguments{
\item{values}{numeric vector}
}
\value{
Logical SW normality test result; message
}
\description{
Shapiro-Wilk test of normality
}
\examples{
set.seed(13)
check_normal(rnorm(50))

}
