% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{convert_genes}
\alias{convert_genes}
\title{Convert list of genes via BioMart}
\usage{
convert_genes(
  genes,
  from_to = c("ensembl_gene_id", "hgnc_symbol"),
  mart = NULL,
  table = FALSE
)
}
\arguments{
\item{genes}{char vector; genes}

\item{from_to}{length 2 char vector; 1st value is current format (ex. \code{ensemb_gene_id}), 2nd value is desired format (ex. \code{hgnc_symbol})}

\item{mart}{biomaRt Mart object}

\item{table}{logical; T to return conversion table}
}
\value{
char vector with converted gene list to desired format
}
\description{
See \code{biomaRT::listAttributes(mart)} for list of all options for \code{from_to}.
}
