% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corner}
\alias{corner}
\title{Like head but just the corner}
\usage{
corner(df, n = 10)
}
\arguments{
\item{df}{dataframe}

\item{n}{integer; number of rows = columns to display}
}
\value{
Top left n x n submatrix of df
}
\description{
Like head but just the corner
}
\examples{
corner(diag(15))
}
