% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_DESeq_PC.R
\name{filter_DESeq_PC}
\alias{filter_DESeq_PC}
\title{Filter DESeq signature to protein coding only}
\usage{
filter_DESeq_PC(
  sig_txt,
  coding_file =
    "/Users/liaoyj/Dropbox/Ovarian Project/log2_coding_expression_datasets/protein-coding_gene.txt"
)
}
\arguments{
\item{sig_txt}{Path to DESeq text file output}

\item{coding_file}{Path to protein-coding gene table w/ "symbol" column}
}
\value{
DESeq text and rnk files subsetted to PC coding only gene
}
\description{
Filter DESeq signature to protein coding only
}
