% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter dataframe by list of (protein coding) genes}
\usage{
filter_genes(
  df,
  genes_filt = Rubrary::get_PC_genes(),
  gene_col = NULL,
  search = FALSE
)
}
\arguments{
\item{df}{dataframe; gene names in \code{gene_col} or in rownames of dataframe}

\item{genes_filt}{char vector; (protein coding) genes to include / filter by}

\item{gene_col}{string; colnames of genes, assumed \code{rownames(df)} if NULL}

\item{search}{logical; \code{TRUE} to use \code{Seurat::UpdateSymbolList} to match gene symbols better}
}
\value{
dataframe with gene rownames filtered to protein coding only
}
\description{
Filter dataframe by list of (protein coding) genes
}
\examples{
df = data.frame(
  gene = c("geneA", "geneB", "geneC", "geneD"),
  sampA = c(1, 2, 3, 4),
  sampB = c(2, 3, 4, 5),
  sampC = c(3, 4, 5, 6)
)
df_filt <- Rubrary::filter_genes(df, genes_filt = c("geneB", "geneC"), gene_col = "gene")

}
