% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{filter_PC_genes}
\alias{filter_PC_genes}
\title{Filter dataframe to protein coding genes}
\usage{
filter_PC_genes(df, PC_genes = Rubrary::get_PC_genes(), search = FALSE)
}
\arguments{
\item{df}{dataframe; rownames must be gene symbols}

\item{PC_genes}{character vector; protein coding genes}

\item{search}{logical; T to use \code{Seurat::UpdateSymbolList} to match gene symbols better}
}
\value{
dataframe with gene rownames filtered to protein coding only
}
\description{
Filter dataframe to protein coding genes
}
