% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{full_join_rownames}
\alias{full_join_rownames}
\title{Full join by rownames}
\usage{
full_join_rownames(df1, df2)
}
\arguments{
\item{df1}{Numerical dataframe with rownames}

\item{df2}{Numerical dataframe with rownames}
}
\value{
Full-joined dataframe by rownames with rownames
}
\description{
Wrapper for dplyr::full_join to merge by rownames.
}
\seealso{
\code{\link[dplyr:mutate-joins]{dplyr::full_join()}}, \code{\link[=left_join_rownames]{left_join_rownames()}}
}
