% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_hypergeo_coexp.R
\name{genecoexp_scatter_hyper}
\alias{genecoexp_scatter_hyper}
\title{Plot pairwise gene coexpression scatter and calculate hypergeometric enrichment p-value between two genelists}
\usage{
genecoexp_scatter_hyper(
  seuobj,
  df_dense = NULL,
  genesA,
  genesB,
  group = NULL,
  savepath,
  ncols = NULL
)
}
\arguments{
\item{seuobj}{Seurat object}

\item{df_dense}{dataframe; dense gene expression matrix from Seurat (cells in rows, genes in columns)}

\item{genesA}{char vector; list of genes}

\item{genesB}{char vector; list of genes}

\item{group}{string; metadata value to group by}

\item{savepath}{string; path to choose from with title sans extension}

\item{ncols}{integer; number of columns in grid}
}
\value{
dataframe of hypergeometric test p-values
}
\description{
Plot pairwise gene coexpression scatter and calculate hypergeometric enrichment p-value between two genelists
}
