% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_gene_desc}
\alias{get_gene_desc}
\title{Get gene description from BioMart}
\usage{
get_gene_desc(
  genes,
  attrs = c("hgnc_symbol", "description"),
  clean_desc = TRUE,
  mart = NULL,
  verbose = FALSE
)
}
\arguments{
\item{genes}{character vector; genes of interest}

\item{attrs}{character vector; valid attributes from \code{biomaRt::listAttributes(mart)}}

\item{clean_desc}{logical; T to remove "[Source: ...]" from description field}

\item{mart}{Mart object; will generate one if not provided}

\item{verbose}{logical; output \code{biomaRt} messages to console}
}
\value{
dataframe with attrs columns, populated from BioMart
}
\description{
Get gene description from BioMart
}
\examples{
genes <- c("POU5F1", "SOX2", "KLF4", "MYC")
get_gene_desc(genes)
}
