% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAsq.R
\name{get_GSEAsq_terms}
\alias{get_GSEAsq_terms}
\title{Get GSEA Squared terms/keywords}
\usage{
get_GSEAsq_terms(
  df_GSEA,
  savename = NULL,
  filt_freq = c(5, 500),
  signlogp_base = 10,
  rep0 = .Machine$double.xmin,
  verbose = TRUE,
  plot = TRUE,
  plot_pval = 1e-05,
  plot_fmt = "png",
  seed = 13
)
}
\arguments{
\item{df_GSEA}{df/string; (path to) GSEA results with \code{pathway} and \code{NES} columns}

\item{savename}{string; path to save output under, no extension}

\item{filt_freq}{num vector; frequency filter for terms, \code{filt_freq[1]} is minimum frequency, \code{filt_freq[2]} is maximum frequency}

\item{signlogp_base}{integer; log base when calculating signed log_base p metric}

\item{rep0}{numeric; value to replace \code{pval == 0} results with, use \code{rep0 = 2.2e-16} (rounded \code{.Machine$double.eps}) to be same as original function}

\item{verbose}{logical; \code{TRUE} to output terms as KS p-value is being calculated}

\item{plot}{logical; \code{TRUE} to output barplot of significant terms}

\item{plot_pval}{numeric; include terms with \code{pval <= plot_pval} in barplot}

\item{plot_fmt}{string; file extension to save plot as}

\item{seed}{numeric; randomization seed}
}
\value{
df w/ terms, frequency, (KS) pval, signed enrichment score, and signlogp
}
\description{
Get GSEA Squared terms/keywords
}
\examples{
library(dplyr)
# Load data
deseq_stats <- setNames(
  airway_deseq_res[,"sign_log_p"],
  airway_deseq_res[,"hgnc_symbol"]
)
pthwys <- GSEA_pathways
# Run (f)gsea
gsea_results <- fgsea::fgsea(
  pathways = pthwys,
  stats = deseq_stats,
  eps = 0.0,
  minSize = 15,
  maxSize  = 500) \%>\%
  arrange(NES)
# Get terms
head(Rubrary::get_GSEAsq_terms(gsea_results, verbose = FALSE))

}
