% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kspval.R
\name{get_kspval}
\alias{get_kspval}
\title{Calculate Kolgomorov-Smirnov enrichment p-value}
\usage{
get_kspval(df, value, group, group2 = NA)
}
\arguments{
\item{df}{dataframe}

\item{value}{string; colname of values to rank by}

\item{group}{string; colname of categorizations}

\item{group2}{string; group of interest}
}
\value{
Kolgomorov-Smirnov enrichment p-value
}
\description{
If no group of interest specified, pulls first condition against others.
}
