% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{get_loadings}
\alias{get_loadings}
\title{Extract loadings from \code{prcomp} object}
\usage{
get_loadings(obj_prcomp)
}
\arguments{
\item{obj_prcomp}{\code{prcomp} object}
}
\value{
Matrix of PCA loadings
}
\description{
Scales the PCA decomposition \emph{eigenvectors} (\code{rotation} output of \code{prcomp}) by the square root of corresponding eigenvalues (\code{sdev} output of \code{prcomp}).
}
\examples{
data(iris)
PCA_iris <- Rubrary::run_PCA(t(iris[,c(1:4)]))
Rubrary::get_loadings(PCA_iris)

}
\seealso{
\code{stats::prcomp}
}
