% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log2_coding_geneexp.R
\name{get_log2_coding_geneexp}
\alias{get_log2_coding_geneexp}
\title{Get log2p1 protein coding gene expression matrix}
\usage{
get_log2_coding_geneexp(exp_file, coding_file, output_filename)
}
\arguments{
\item{exp_file}{Gene exp data; samples as columns, genes as rows}

\item{coding_file}{Protein coding genes mtx; "symbol" column}

\item{output_filename}{Filename to save output as}
}
\value{
txt (tsv) with log2p1 transformed protein coding only gene expression
}
\description{
Get log2p1 protein coding gene expression matrix
}
