% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log2_coding_geneexp.R
\name{get_log_coding_geneexp}
\alias{get_log_coding_geneexp}
\title{Get log(x+1) protein coding gene expression matrix}
\usage{
get_log_coding_geneexp(
  exp,
  base = 2,
  pc_genes = Rubrary::get_PC_genes(),
  savename
)
}
\arguments{
\item{exp}{df/string; samples as columns, genes as rows}

\item{base}{integer; log transformation base}

\item{pc_genes}{char vector; list of coding genes to filter by}

\item{savename}{string; filename to save output as}
}
\value{
Log(x+1) transformed protein coding only gene expression dataframe
}
\description{
Get log(x+1) protein coding gene expression matrix
}
