% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DESeq_waterfall_hl.R
\name{get_n}
\alias{get_n}
\title{Get number of TF MRA signature genes to highlight}
\usage{
get_n(signature, pos = T, sig = F)
}
\arguments{
\item{signature}{df; Has "pvalue" and "NES" columns}

\item{pos}{logical; "T" for positive NES}

\item{sig}{logical; "T" for p < 0.05}
}
\value{
integer with n number of genes to highlight
}
\description{
Get number of TF MRA signature genes to highlight
}
