% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{get_PC_genes}
\alias{get_PC_genes}
\title{Retrieve protein coding genes from Ensembl BioMart}
\usage{
get_PC_genes(mart = NULL)
}
\arguments{
\item{mart}{biomaRt Mart object}
}
\value{
character vector of protein coding genes
}
\description{
Retrieve protein coding genes from Ensembl BioMart
}
\examples{
# head(Rubrary::get_PC_genes())

}
