% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSEA_pathways}
\alias{GSEA_pathways}
\title{Pathways from \code{msigdbr} to use with \code{fgsea}}
\format{
\subsection{\code{GSEA_pathways}}{

A list with 10,481 gene sets
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/human/collection_details.jsp}
}
\usage{
GSEA_pathways
}
\description{
Broad set of pathways extracted from \code{msigdbr} database and formatted as a list, with names being gene set names and values being vectors of genes in a given gene set.
}
\details{
Includes \href{https://www.gsea-msigdb.org/gsea/msigdb/human/collection_details.jsp}{Human MSigDB Collections}:
\itemize{
\item C2CP: curated gene sets, canonical pathways
\item C5GO: ontology gene sets, Gene Ontology gene sets
\item BP: GO Biological Process ontology
\item CC: GO Cellular Component ontology
\item MF: GO Molecular Function ontology
\item H: hallmark gene sets
}
}
\keyword{datasets}
