% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inst_pkg}
\alias{inst_pkg}
\title{Prompt user to install package}
\usage{
inst_pkg(pkg)
}
\arguments{
\item{pkg}{string; package name (CRAN or Bioconductor)}
}
\value{
None
}
\description{
Helper function for "use_pkg"
}
