% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left_join_rownames}
\alias{left_join_rownames}
\title{Left join by rownames}
\usage{
left_join_rownames(df1, df2)
}
\arguments{
\item{df1}{Numerical dataframe with rownames}

\item{df2}{Numerical dataframe with rownames}
}
\value{
Left-joined dataframe by rownames with rownames
}
\description{
Wrapper for dplyr::left_join to merge by rownames. Both dataframes should have matching rownames
}
