% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq.R
\name{make_pconly}
\alias{make_pconly}
\title{Filter signature to PC coding only by matching with genes in log2UQcounts}
\usage{
make_pconly(sig_path)
}
\arguments{
\item{sig_path}{Path to DESeq signature}
}
\value{
Text and rnk files subsetted to PC coding only genes
}
\description{
Filter signature to PC coding only by matching with genes in log2UQcounts
}
