% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgenes.R
\name{output_DESeq}
\alias{output_DESeq}
\title{Output DESeq results as dataframe with signed log2 p metric}
\usage{
output_DESeq(DE_results, savename = NULL, rank = FALSE)
}
\arguments{
\item{DE_results}{DESeq results object}

\item{savename}{string; filepath to save table under (no ext.)}

\item{rank}{logical; T to output .rnk file compatible with Java GSEA app}
}
\value{
Dataframe with DESeq results
}
\description{
Filters out all \code{sign_log_p == NA} rows.
}
