% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_hypergeo_coexp.R
\name{phyper_df}
\alias{phyper_df}
\title{Hypergeometric test, 0 vs. non-0, on columns of a dataframe}
\usage{
phyper_df(df, A, B)
}
\arguments{
\item{df}{dataframe; includes columns "A" and "B"}

\item{A}{string; colname for values A}

\item{B}{string; colname for values B}
}
\value{
single dataframe row with A&B names, phyper inputs, pvalues
}
\description{
Originally used for scRNA gene co-expression analysis
}
