% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot_paired.R
\name{plot_boxplot_paired}
\alias{plot_boxplot_paired}
\title{plot_boxplot_paired}
\usage{
plot_boxplot_paired(sig, sig_paired_df, batch)
}
\arguments{
\item{sig}{Column name of specific signature}

\item{sig_paired_df}{Dataframe with all signatures}

\item{batch}{Plot subtitle}
}
\value{
Paired boxplot separated by type
}
\description{
plot_boxplot_paired
}
