% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgenes.R
\name{plot_DEgene_boxplot}
\alias{plot_DEgene_boxplot}
\title{Plot gene expression boxplots of selected genes (from DESeq signature)}
\usage{
plot_DEgene_boxplot(
  data,
  anno,
  anno_sig,
  anno_type = NULL,
  sig_DE,
  genes = NULL,
  metric = "sign_log_p",
  pval = FALSE,
  genes_n = 6,
  colors = NULL,
  title = NULL,
  savename = NULL,
  height = 8,
  width = 8
)
}
\arguments{
\item{data}{df/string; (path to) numeric gene expression matrix, rownames = genes, colnames = samples}

\item{anno}{df/string; (path to) annotation table, rownames = samples, columns = metadata}

\item{anno_sig}{string; colname in \code{anno} to be used as breaks for x-axis}

\item{anno_type}{string; colname in \code{anno} to be used to color jitter points by}

\item{sig_DE}{df/string, optional; (path to) DE(seq) gene results table}

\item{genes}{char vector; list of genes to plot, must be present in \code{data}}

\item{metric}{string; colname in \code{sig_DE} to rank genes by for auto gene selection if \code{genes} not provided and for appending to gene name in parens}

\item{pval}{logical; use \code{t.test} pvalue annotation via \code{ggsignif::stat_signif}}

\item{genes_n}{integer; if \code{genes} not provided, get top/bottom \code{genes_n}}

\item{colors}{char vector; jitter point colors, length == # of unique \code{anno_type}s}

\item{title}{string; plot title}

\item{savename}{string; file path to save plot under}

\item{height}{numeric; plot height}

\item{width}{numeric; plot width}
}
\value{
ggplot object, with facet-wrapped boxplot
}
\description{
Great for sanity checking, to make sure that differentially expressed genes as reported through DE gene analysis metrics (such as DESeq2) are reflected in gene expression data per sample.
}
\details{
Providing DE(seq) gene signature (\code{sig_DE}) is optional; if not provided, \code{genes} list must be defined. If DE gene info \emph{is} provided, each subplot's title will have the \code{metric} value appended in parenthesis.
}
