% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq.R
\name{plot_DESeq_boxplot}
\alias{plot_DESeq_boxplot}
\title{plot_DESeq_boxplot}
\usage{
plot_DESeq_boxplot(sig, data, anno, batch, merged = T)
}
\arguments{
\item{sig}{Signature}

\item{data}{Gene expression data}

\item{anno}{Annotation file for samples}

\item{batch}{Descriptive word/phrase appended to output path}

\item{merged}{If dataframe is made from combined datasets}
}
\value{
Paired boxplots comparing pre-post for top/bot 5 DESeq genes, plotting gene expression data
}
\description{
plot_DESeq_boxplot
}
