% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq.R
\name{plot_DESeq_scatter}
\alias{plot_DESeq_scatter}
\title{plot_DESeq_scatter}
\usage{
plot_DESeq_scatter(
  sig1path,
  sig2path,
  batch = "",
  xlab = "Old DESeq signed log p",
  ylab = "Updated DESeq signed log p",
  title = "",
  desc = "",
  rank = F
)
}
\arguments{
\item{sig1path}{Path to 1st signature (x-axis)}

\item{sig2path}{Path to 2nd signature (y-axis)}

\item{batch}{Descriptive word/phrase appended to output path}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{title}{Plot title}

\item{desc}{Plot subtitle}

\item{rank}{Plot rank scatter instead of slogp/raw values}
}
\value{
Plot comparing 1 sig to another
}
\description{
plot_DESeq_scatter
}
