% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DESeq_volcano.R
\name{plot_DESeq_volcano}
\alias{plot_DESeq_volcano}
\title{Plot DESeq volcano}
\usage{
plot_DESeq_volcano(
  df_deseq,
  pCutoff = 5e-04,
  logFCcutoff = 1,
  title = "DE Genes",
  subtitle = "",
  savename = NA
)
}
\arguments{
\item{df_deseq}{dataframe; DESeq output read as df}

\item{pCutoff}{numeric; value to set p value cutoff}

\item{logFCcutoff}{numeric; value to set log fold change cutoff}

\item{title}{string; plot title}

\item{subtitle}{string; plot subtitle}

\item{savename}{string; full name of file to save under}
}
\value{
DE genes volcano plot
}
\description{
Plot DESeq volcano
}
