% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DESeq_waterfall_hl.R
\name{plot_DESeq_waterfall_hl}
\alias{plot_DESeq_waterfall_hl}
\title{Plot DESeq waterfall with genes highlighted}
\usage{
plot_DESeq_waterfall_hl(
  sig,
  sig_hl,
  topn = 200,
  rev = F,
  toplabel = "Top SCN",
  title = "DE Genes",
  subtitle = NA,
  save = F,
  savename = "DESeq_slogp_WF_hl.png"
)
}
\arguments{
\item{sig}{df; DESeq signature with "sign_log_p" and "gene" columns}

\item{sig_hl}{df; ranked signature with genes in column 1}

\item{topn}{integer; number of genes from sig_hl to highlight}

\item{rev}{logical; reverse to bottom n genes from sig_hl instead of top}

\item{toplabel}{string; name of highlight group}

\item{title}{string; plot title}

\item{subtitle}{string; plot subtitle}

\item{save}{logical; "T" to save}

\item{savename}{string; name to save plot under}
}
\value{
Waterfall plot with n genes from sig_hl highlighted, ks p value
}
\description{
Revamp of Favour's original function
}
