% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dim_set.R
\name{plot_dim_set}
\alias{plot_dim_set}
\title{Plot Seurat DimPlot grid set}
\usage{
plot_dim_set(
  sobj,
  reduction = "umap",
  set,
  ncol = 3,
  guides = NULL,
  savename = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{sobj}{Seurat object}

\item{reduction}{c("pca", "tsne", "umap); must be already run in sobj}

\item{set}{dataframe; col1 metadata name, col2 subplot description, col3 (opt) logical T/F for label}

\item{ncol}{number of columns in grid}

\item{guides}{string; specify how guides should be treated in layout (?patchwork::wrap_plots)}

\item{savename}{string; filepath to save figure under}

\item{width}{integer; width of plot}

\item{height}{integer; height of plot}
}
\value{
Patchwork grid of Seurat DimPlots
}
\description{
Plot Seurat DimPlot grid set
}
