% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{plot_dimplot_grid}
\alias{plot_dimplot_grid}
\title{Plot Seurat DimPlot grid set}
\usage{
plot_dimplot_grid(
  sobj,
  reduction = c("umap", "pca", "tsne"),
  set,
  ncol = 3,
  guides = NULL,
  redtitle = FALSE,
  gridtitle = NULL,
  gridsubtitle = NULL,
  cols_num = c("blue", "red"),
  savename = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{sobj}{Seurat object}

\item{reduction}{c("pca", "tsne", "umap); must be already present in sobj}

\item{set}{dataframe or vector; col1 metadata name, col2 (opt) subplot description, col3 (opt) logical T/F for label}

\item{ncol}{number of columns in grid}

\item{guides}{string; specify how guides should be treated in layout (?patchwork::wrap_plots)}

\item{redtitle}{logical; \code{TRUE} to prepend reduction method in subplot title}

\item{gridtitle}{string; title for overall dimplot grid}

\item{gridsubtitle}{string; subtitle for overall dimplot grid}

\item{cols_num}{char vector; two colors for low numeric legend range to high numeric legend range}

\item{savename}{string; filepath to save figure under}

\item{width}{integer; plot width, if unspecified will try to calculate}

\item{height}{integer; plot height, if unspecified will try to calculate}
}
\value{
Patchwork grid of Seurat DimPlots
}
\description{
Plot Seurat DimPlot grid set
}
