% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot distribution / normality check}
\usage{
plot_distribution(
  df = NA,
  value,
  swtest = FALSE,
  title = "Distribution",
  xlab = "Value",
  save = FALSE,
  savename = "Distribution.png"
)
}
\arguments{
\item{df}{dataframe}

\item{value}{string; colname of values}

\item{swtest}{logical; perform SW test for normality?}

\item{title}{string; plot title}

\item{xlab}{string; x-axis label}

\item{save}{logical; save as PNG?}

\item{savename}{string; file to save figure as}
}
\value{
Plot of distribution with median + mean indicated
}
\description{
Plot distribution / normality check
}
\examples{
set.seed(13)
df_norm <- data.frame(Value = rnorm(50))
plot_distribution(df_norm, value = "Value", title = "Normal", swtest = TRUE)

df_uni <- data.frame(Value = 1:100)
plot_distribution(df_uni, value = "Value", title = "Uniform", swtest = TRUE)

}
