% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{plot_featurescatter}
\alias{plot_featurescatter}
\title{Plot FeatureScatter wrapper}
\usage{
plot_featurescatter(
  sobj,
  feat1,
  feat2,
  group = "sample_id",
  title = NULL,
  savename = NULL,
  corr_method = c("pearson", "kendall", "spearman"),
  height = 7,
  width = 9
)
}
\arguments{
\item{sobj}{Seurat object}

\item{feat1}{string; 1st feature to plot (x-axis)}

\item{feat2}{string; 2nd feature to plot (y-axis)}

\item{group}{string; metadata colname to group by}

\item{title}{string; plot title}

\item{savename}{string; filepath to save plot under}

\item{corr_method}{string; \code{stats::cor} method for correlation calculation}

\item{height}{numeric; plot height}

\item{width}{numeric; plot width}
}
\value{
FeatureScatter ggplot output
}
\description{
Wrapper for \code{Seurat::FeatureScatter} with a few extras like saving options, plot title, and correlation in the plot subtitle
}
