% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{plot_GSEA_barplot}
\alias{plot_GSEA_barplot}
\title{Plot GSEA NES barplot}
\usage{
plot_GSEA_barplot(
  gsea_res,
  gsea_name = "GSEA",
  gsea_pws = NULL,
  n_pws = 5,
  pw_format = FALSE,
  pw_split = FALSE,
  pw_source = TRUE,
  pw_ignore = NULL,
  pw_size = 5,
  gsea2_res = NULL,
  gsea2_name = NULL,
  order2 = FALSE,
  ptrn2 = "none",
  NES_cutoff = NULL,
  sig_cutoff = NULL,
  colors = c("firebrick", "darkblue"),
  title = NULL,
  savename = NULL,
  width = 10,
  height = NULL
)
}
\arguments{
\item{gsea_res}{df/string; (path to) GSEA results w/ \code{pathway} and \code{NES} columns}

\item{gsea_name}{string; description of GSEA results}

\item{gsea_pws}{char vector; pathways in GSEA results to plot}

\item{n_pws}{integer; if no pathways provided, top/bottom n pathways (ordered by NES) to plot}

\item{pw_format}{logical; clean up pathway names?}

\item{pw_split}{logical; T to split pathway names into multiple lines}

\item{pw_source}{logical; T to append pathway source in parenthesis}

\item{pw_ignore}{char vector; list of terms to ignore for name formatting}

\item{pw_size}{numeric; pathway name text size}

\item{gsea2_res}{df/string; (path to) 2nd GSEA results w/ \code{pathway} and \code{NES} columns}

\item{gsea2_name}{string; description of 2nd GSEA results}

\item{order2}{logical; order pathways by NES of 2nd GSEA results?}

\item{ptrn2}{string; \code{ggpattern} pattern arg: 'stripe', 'crosshatch', 'point', 'circle', 'none'}

\item{NES_cutoff}{numeric; value to draw NES cutoff line at}

\item{sig_cutoff}{vector; \code{sig_cutoff[1]} is colname of significance values ("pval" or "padj"), \code{sig_cutoff[2]} is numeric cutoff value: if higher than this value, alpha will be 0.5. Ex. \code{sig_cutoff = c("pval", 0.05)} or \code{sig_cutoff = c("padj", 0.1)}}

\item{colors}{vector; \code{colors[1]} for positive bar color, \code{colors[2]} for negative bar color}

\item{title}{string; plot title}

\item{savename}{string; file path to save plot under}

\item{width}{numeric; plot width}

\item{height}{numeric; plot height}
}
\value{
GSEA NES barplot as \code{ggplot} object
}
\description{
Can plot barplot of NES of pathways from one GSEA result, or compare NES for chosen pathways between two GSEA results.
}
\details{
To extend margins and accomodate long pathway names, add \code{ggplot2::theme(plot.margin = unit(c(top, right, bottom, left), units))} to the ggplot object.
}
