% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSEA_pathway.R
\name{plot_GSEA_batch}
\alias{plot_GSEA_batch}
\title{\code{plot_GSEA_pathway} that works nicely with \code{lapply}}
\usage{
plot_GSEA_batch(
  path_name,
  pthwys,
  sig_name,
  sig,
  rankcol,
  rankcol_name,
  hllab = "Highlight",
  lowlab = "Low",
  highlab = "High",
  savedir = "./",
  label = length(pthwys[[path_name]]) < 50,
  hightolow = FALSE
)
}
\arguments{
\item{path_name}{string; name of pathway}

\item{pthwys}{named list; key = geneset name, values = char vector of genes in geneset}

\item{sig_name}{string; name of signature}

\item{sig}{dataframe; signature}

\item{rankcol}{string; colname of values to rank by}

\item{rankcol_name}{string; descriptor of rankcol}

\item{hllab}{string; descriptor of highlighted genes}

\item{lowlab}{string; label for low rankcol values}

\item{highlab}{string; label for high rankcol values}

\item{savedir}{string; directory path for saving plot}

\item{label}{logical; T to label highlighted genes}

\item{hightolow}{logical; T for high values on left, low on right}
}
\value{
nothing; will save GSEA enrichment plots in savedir
}
\description{
\code{plot_GSEA_pathway} that works nicely with \code{lapply}
}
