% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{plot_GSEA_pathway_batch}
\alias{plot_GSEA_pathway_batch}
\title{\code{plot_GSEA_pathway} wrapper that works nicely with \code{lapply}}
\usage{
plot_GSEA_pathway_batch(
  path_name,
  pthwys,
  sig,
  genecol = "gene",
  rankcol,
  rankcol_name = rankcol,
  hllab = "Pathway genes",
  hightolow = FALSE,
  format_name = TRUE,
  ignore_name = NULL,
  lab_low = NULL,
  lab_high = NULL,
  legendpos = c(0.5, 0.2),
  hl_color = "firebrick3",
  label = length(pthwys[[path_name]]) < 50,
  savedir = NULL,
  sig_name = ""
)
}
\arguments{
\item{path_name}{string; name of pathway}

\item{pthwys}{named list; key = geneset name, values = char vector of genes in geneset}

\item{sig}{dataframe; signature}

\item{genecol}{string; colname of gene names in \code{sig}}

\item{rankcol}{string; colname of values to rank by}

\item{rankcol_name}{string; descriptor of rankcol}

\item{hllab}{string; descriptor of highlighted genes}

\item{hightolow}{logical; T for high values on left, low on right}

\item{format_name}{logical; T to run pathway name through \code{Rubrary::format_GSEA_name}}

\item{ignore_name}{char vector; phrases to exclude title case for in \code{Rubrary::format_GSEA_name}}

\item{lab_low}{string; label for low rankcol values}

\item{lab_high}{string; label for high rankcol values}

\item{legendpos}{vector; value btwn 0-1 as legend coordinates (ggplot2 legend.position)}

\item{hl_color}{string; color for highlight}

\item{label}{logical; T to label points in plot}

\item{savedir}{string; directory path for saving plot}

\item{sig_name}{string; name of signature to append to save path}
}
\value{
Gene set enrichment plot as ggplot object
}
\description{
\code{plot_GSEA_pathway} wrapper that works nicely with \code{lapply}
}
\examples{
airway_deseq = Rubrary::airway_deseq_res
pathways <- Rubrary::GSEA_pathways
pws_plot <- c("HALLMARK_ADIPOGENESIS", "GOCC_POSTSYNAPTIC_MEMBRANE")
lapply(
  pws_plot,
  Rubrary::plot_GSEA_pathway_batch,
  pthwys = pathways,
  sig = airway_deseq,
  genecol = "hgnc_symbol",
  rankcol = "sign_log_p",
  rankcol_name = "Sign log p value",
  lab_high = "\U2191 in treated\n\U2193 in untreated",
  lab_low = "\U2191 in untreated\n\U2193 in treated",
)

}
