% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{plot_GSEAsq_density}
\alias{plot_GSEAsq_density}
\title{Plot categorical percentile rank comparison between two GSEA squared signatures}
\usage{
plot_GSEAsq_density(
  GSEAsq_df1,
  GSEAsq_df2,
  name1,
  name2,
  title = "GSEA Squared",
  colors = c("firebrick", "lightslateblue"),
  rug = TRUE,
  savename = NULL,
  height = 10,
  width = 10
)
}
\arguments{
\item{GSEAsq_df1}{df/string; "gsea_squared-df.txt" df or path for GSEAsq result 1}

\item{GSEAsq_df2}{df/string; "gsea_squared-df.txt" df or path for GSEAsq result 2}

\item{name1}{string; descriptor for GSEAsq result 1}

\item{name2}{string; descriptor for GSEAsq result 2}

\item{title}{string; overall plot title}

\item{colors}{vector; first color to signify result 1, second for result 2}

\item{rug}{logical; TRUE for rugplot below density plot}

\item{savename}{string; filepath to save plot under}

\item{height}{numeric; plot height}

\item{width}{numeric; plot width}
}
\value{
facet wrapped ggplot object
}
\description{
Plot categorical percentile rank comparison between two GSEA squared signatures
}
