% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA with annotation}
\usage{
plot_PCA(
  df_pca,
  anno = NULL,
  PCx = "PC1",
  PCy = "PC2",
  type = c("Scores", "Loadings"),
  annoname = "Sample",
  annolabel = annoname,
  label = FALSE,
  annotype = "Type",
  annotype2 = NULL,
  ellipse = FALSE,
  ks_pval = c("none", "caption", "grob"),
  highlight = NULL,
  colors = NULL,
  col_midpt = 0,
  title = NULL,
  subtitle = NULL,
  density = FALSE,
  savename = NULL,
  width = 8,
  height = 8
)
}
\arguments{
\item{df_pca}{string or \code{prcomp} obj; (path to) PCA output}

\item{anno}{string or df; Annotation info for DF}

\item{PCx}{string; Component on x-axis}

\item{PCy}{string; Component on y-axis}

\item{type}{c("Score", "Loading")}

\item{annoname}{string; Colname in \code{anno} matching point name}

\item{annolabel}{string; Colname in \code{anno} to label points by, defaults to \code{annoname}}

\item{label}{logical; T to label points}

\item{annotype}{string; Colname in \code{anno} with info to color by}

\item{annotype2}{string; Colname in \code{anno} with info to change shape by}

\item{ellipse}{logical; Draw \code{ggplot2::stat_ellipse} data ellipse w/ default params - this is NOT a confidence ellipse}

\item{ks_pval}{c("none", "caption", "grob"); Display ks-pvalue as caption or grob}

\item{highlight}{char vector; Specific points to shape differently & label}

\item{colors}{char vector; For discrete \code{annotype}, length should be number of unique \code{annotype}s. For continuous \code{annotype}, can either be length 2 where \code{colors[1]} is low and \code{colors[2]} is high or length 3 diverging colorscale where \code{colors[1]} = low, \code{colors[2]} = mid, \code{colors[3]} = high.}

\item{col_midpt}{numeric; For continuous \code{length(colors) == 3} \code{scale_color_gradientn} usage only}

\item{title}{string; Plot title}

\item{subtitle}{string; Subtitle for plot}

\item{density}{logical; Show density plot along both axes; requires group annotations to be provided}

\item{savename}{string; File path to save plot under}

\item{width}{numeric; Saved plot width}

\item{height}{numeric; Saved plot height}
}
\value{
PCA output plotted with annotation
}
\description{
Plot PCA with annotation
}
\examples{
data(iris)
iris$Sample = rownames(iris)
PCA_iris <- Rubrary::run_PCA(t(iris[,c(1:4)]))
# Scores
Rubrary::plot_PCA(
  df_pca = PCA_iris,
  anno = iris[,c("Sample", "Species")],
  annoname = "Sample", annotype = "Species",
  title = "Iris PCA Scores by Species",
  subtitle = "Centered & scaled",
  ellipse = TRUE
)
# Loadings
Rubrary::plot_PCA(
  df_pca = PCA_iris,
  type = "Loadings",
  title = "Iris PCA Loadings",
  subtitle = "Centered & scaled",
  label = TRUE
)

}
