% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot_PCA_biplot}
\alias{plot_PCA_biplot}
\title{Plot "proper" PCA biplot}
\usage{
plot_PCA_biplot(
  obj,
  PCx = "PC1",
  PCy = "PC2",
  anno = NULL,
  annoname = "Sample",
  annotype = "Batch",
  label = c("Both", "Loadings", "Scores", "None"),
  colors = NULL,
  col_load = "firebrick",
  title = NULL,
  ellipse = FALSE,
  savename = NULL,
  height = 8,
  width = 8
)
}
\arguments{
\item{obj}{\code{prcomp} object}

\item{PCx}{string; Component on x-axis}

\item{PCy}{string; Component on y-axis}

\item{anno}{df; Annotation info for observations}

\item{annoname}{string; Colname in \code{anno} matching data points}

\item{annotype}{string; Colname in \code{anno} for desired coloring}

\item{label}{c("Both", "Loadings", "Scores", "None"); what points to label}

\item{colors}{char vector; Length should be number of unique \code{annotype}s}

\item{col_load}{string; Color for loading arrow segments}

\item{title}{string; Plot title}

\item{ellipse}{logical; Draw \code{ggplot2::stat_ellipse} data ellipse w/ default params - this is NOT a confidence ellipse}

\item{savename}{string; File path to save plot under}

\item{height}{numeric; Saved plot height}

\item{width}{numeric; Saved plot width}
}
\value{
Biplot as \code{ggplot} object
}
\description{
For PCA SVD \eqn{X = USV^T}, uses standardized principal components for data points (\eqn{\bf{U}\sqrt{n - 1}}) and loadings (\eqn{\bf{VS}/\sqrt{n-1}}) and plots onto the same scale - a "proper" PCA biplot according to \href{https://stats.stackexchange.com/questions/141085/positioning-the-arrows-on-a-pca-biplot/141531}{this Stack Exchange thread} citing the \href{https://doi.org/10.2307/2334381}{Gabriel 1971 paper on PCA biplots}.
}
\examples{
data(iris)
iris$Sample = rownames(iris)
PCA_iris <- Rubrary::run_PCA(t(iris[,c(1:4)]),
  center = TRUE, scale = FALSE, screeplot = FALSE)
Rubrary::plot_PCA_biplot(
  obj = PCA_iris,
  anno = iris[,c("Sample", "Species")],
  annoname = "Sample", annotype = "Species",
  label = "Loadings", ellipse = TRUE, title = "Iris PCA Biplot")
}
