% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot_PCA_matrix}
\alias{plot_PCA_matrix}
\title{Plot PCA matrix via \code{GGally::ggpairs}}
\usage{
plot_PCA_matrix(
  df_pca,
  PCs = c(1:3),
  anno = NULL,
  annoname = "Sample",
  annotype = "Type",
  colors = NULL,
  title = NULL,
  savename = NULL,
  width = 8,
  height = 8
)
}
\arguments{
\item{df_pca}{string/\code{prcomp} obj; (path to) PCA output with sample names in first column}

\item{PCs}{num vector; list of numeric PCs to plot (ex. \code{c(1:3)} for first 3 PCs)}

\item{anno}{string/df; Annotation info for DF}

\item{annoname}{string; Colname in \code{anno} matching point name}

\item{annotype}{string; Colname in \code{anno} with info to color by}

\item{colors}{char vector; For discrete \code{annotype}, length should be number of unique \code{annotype}s. For continuous \code{annotype}, can either be length 2 where \code{colors[1]} is low and \code{colors[2]} is high or length 3 diverging colorscale where \code{colors[1]} = low, \code{colors[2]} = mid, \code{colors[3]} = high.}

\item{title}{string; Plot title}

\item{savename}{string; File path to save plot under}

\item{width}{numeric; Saved plot width}

\item{height}{numeric; Saved plot height}
}
\value{
\code{ggmatrix} object with PCA scatter plot matrix
}
\description{
Plot PCA matrix via \code{GGally::ggpairs}
}
\examples{
data(iris)
iris$Sample = rownames(iris)
plot_PCA_matrix(
  df_pca = Rubrary::run_PCA(t(iris[,c(1:4)]), screeplot = FALSE),
  PCs = c(1:3),
  anno = iris[,c("Sample", "Species")],
  annoname = "Sample",
  annotype = "Species",
  title = "Iris PCA"
)
}
