% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_compare.R
\name{plot_scatter_compare}
\alias{plot_scatter_compare}
\title{plot_scatter_compare}
\usage{
plot_scatter_compare(
  set1path,
  set2path,
  set1lab = "Set 1",
  set2lab = "Set 2",
  rank = F,
  save = F,
  title = paste0(set1lab, " vs. ", set2lab),
  subtitle = "",
  outname = paste0(dirname(set1path), "/", set1lab, "vs", set2lab, ".png")
)
}
\arguments{
\item{set1path}{Path to 1st set (x axis)}

\item{set2path}{Path to 2nd set (y axis)}

\item{set1lab}{Axis label for 1st set (x)}

\item{set2lab}{Axis label for 2nd set (y)}

\item{rank}{Convert sign log p values to rank}

\item{save}{Save as .png}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{outname}{File output}
}
\value{
Plot
}
\description{
plot_scatter_compare
}
