% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SCN_KaplanMeier.R
\name{plot_SCN_KaplanMeier}
\alias{plot_SCN_KaplanMeier}
\title{Plot SCN Kaplan Meier Survival Analysis}
\usage{
plot_SCN_KaplanMeier(
  df,
  xlab = "Time",
  title = "Survival Analysis",
  SCN = NA,
  save = T,
  filename = paste0(gsub(" ", "_", title, fixed = TRUE), ".png")
)
}
\arguments{
\item{df}{dataframe w/ columns "OS" (overall survival censor), "OS.time" (overall survival time)}

\item{xlab}{x-axis label}

\item{title}{plot title}

\item{SCN}{SCN threshold to designate SCN-like status if not already in DF}

\item{save}{save plot as png}

\item{filename}{filename to save as}
}
\value{
Kaplan-Meier overall survival analysis plot comparing SCN-like status
}
\description{
Plot SCN Kaplan Meier Survival Analysis
}
