% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_screeplot.R
\name{plot_screeplot}
\alias{plot_screeplot}
\title{Plot screeplot from prcomp}
\usage{
plot_screeplot(
  obj_prcomp,
  npcs = ncol(obj_prcomp$x),
  label = FALSE,
  cum_var_exp = 80,
  savename = NULL
)
}
\arguments{
\item{obj_prcomp}{prcomp function output object}

\item{npcs}{integer; # of PCs to include in plot}

\item{label}{logical; T for \% values at points}

\item{cum_var_exp}{integer; c(1:100), cumulative variance explained threshold}

\item{savename}{string; filepath to save PNG under}
}
\value{
Screeplot
}
\description{
Plot screeplot from prcomp
}
