% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall_hl.R
\name{plot_waterfall_hl}
\alias{plot_waterfall_hl}
\title{Plot ranked waterfall with genes highlighted}
\usage{
plot_waterfall_hl(
  sig,
  rankcol = "sign_log_p",
  sig_hl,
  topn = 200,
  rev = F,
  toplabel = "Top SCN",
  label = F,
  ylab = "DESeq Signed log p-values",
  wf_pos = "Post-Therapy",
  wf_neg = "Pre-Therapy",
  title = "DE Genes",
  subtitle = NA,
  pval = T,
  density = T,
  save = F,
  savename = "DESeq_slogp_WF_hl.png"
)
}
\arguments{
\item{sig}{df; Ranked signature with "rankcol" and "gene" columns}

\item{rankcol}{string; column to rank sig df by}

\item{sig_hl}{df; ranked signature with genes in column 1}

\item{topn}{integer; number of genes from sig_hl to highlight}

\item{rev}{logical; reverse to bottom n genes from sig_hl instead of top}

\item{toplabel}{string; name of highlight group}

\item{label}{logical or char vector; if vector, should match length of nrow(sig_hl)}

\item{ylab}{string; rankcol axis label}

\item{wf_pos}{string; label for positive side of waterfall}

\item{wf_neg}{string; label for negative side of waterfall}

\item{title}{string; plot title}

\item{subtitle}{string; plot subtitle}

\item{pval}{logical; include KS p value?}

\item{density}{logcal; include density plot on right?}

\item{save}{logical; "T" to save}

\item{savename}{string; name to save plot under}
}
\value{
Waterfall plot with n genes from sig_hl highlighted, ks p value
}
\description{
Revamp of Favour's original function
}
