% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_norm.R
\name{quantile_norm}
\alias{quantile_norm}
\title{Quantile normalization, upper quartile by default}
\usage{
quantile_norm(vals, q = 75, target = 1000, min = 1, perl = FALSE)
}
\arguments{
\item{vals}{numeric df/vector}

\item{q}{numeric; quartile to normalize by (ex. \code{q = 75} is 75th quantile / upper quartile)}

\item{target}{numeric; value to scale by}

\item{min}{numeric; filter to values above \code{min} for quantile normalization}

\item{perl}{logical; \code{TRUE} to match output of \code{quartile_norm.pl} commonly used in Graeber Lab RNA-seq analysis}
}
\value{
Quantile normalized numeric vector or dataframe
}
\description{
In RNA-seq normalization, upper quartile normalization = divide each read count by 75th quantile of read counts in its sample.
}
