% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reload_pkg}
\alias{reload_pkg}
\title{Reload package by installation path}
\usage{
reload_pkg(pkg = "Rubrary")
}
\arguments{
\item{pkg}{string; name of package}
}
\value{
Reloads pkg
}
\description{
Reload package by installation path
}
