% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rread}
\alias{rread}
\title{Wrapper for \code{data.table::fread} implementing \code{row.names} functionality}
\usage{
rread(input, row.names = 0, to_df = TRUE)
}
\arguments{
\item{input}{string; \code{fread} input}

\item{row.names}{integer; column to use as row names, 0 for none}

\item{to_df}{logical; T to convert output to data.frame}
}
\value{
Table read in from input
}
\description{
Default file reading function in Rubrary functions because data is often in large tables and \code{read.delim} isn't optimized for high dimensionality. If file extension is \code{xlsx}, will attempt to use \code{openxlsx::read.xlsx} instead. Setting \code{row.names != 0} will result in output being a \code{data.frame}.
}
\examples{
glab_Beltran_2016 <- "https://raw.githubusercontent.com/graeberlab-ucla/glab.library/master/vignettes/PCA_tutorial/Beltran_2016_rsem_genes_upper_norm_counts_coding_log2.txt"
df <- Rubrary::rread(glab_Beltran_2016, row.names = 1, to_df = FALSE)
Rubrary::corner(df, 5)
}
