% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq_old.R
\name{run_DESeq}
\alias{run_DESeq}
\title{Run DESeq}
\usage{
run_DESeq(mtx_rawcts, mtx_annotation, savename = NULL, merged = F, shrink = F)
}
\arguments{
\item{mtx_rawcts}{matrix; numeric matrix of raw counts}

\item{mtx_annotation}{matrix; samples as rownames, cols "Subject" & "Condition" (& "Batch" if merged)}

\item{savename}{string; filename to save outputs under (no ext.)}

\item{merged}{logical; T if dataset has two batches merged}

\item{shrink}{logical; apply lfcShrink "ashr" to results (NOT IMPLEMENTED)}
}
\value{
DESeq results in dataframe
}
\description{
Not generalizable to non OV project related datasets yet...
}
