% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAsq.R
\name{run_GSEA_squared}
\alias{run_GSEA_squared}
\title{Run GSEA Squared on GSEA results}
\usage{
run_GSEA_squared(
  df_GSEA,
  categories,
  cat_terms,
  rep0 = 2.2e-16,
  signlogp_base = 10,
  get_terms = FALSE,
  terms_filt_freq = c(5, 500),
  plot_type = c("jitter", "density"),
  plot_pval = TRUE,
  title = NULL,
  cat_colors = NULL,
  savename = NULL,
  plot_fmt = "png",
  height = 8,
  width = 8,
  seed = 13,
  verbose = TRUE
)
}
\arguments{
\item{df_GSEA}{df/string; (path to) GSEA results with \code{pathway} and \code{NES} columns}

\item{categories}{char vector; list of category names}

\item{cat_terms}{char vector; list of category terms with each element being keywords separated by "\code{|}", ex. "CELL_CYCLE|MITOTIC|DNA_REPLICATION"}

\item{rep0}{numeric; value to replace \code{pval == 0} results with, use \code{rep0 = 2.2e-16} (rounded \code{.Machine$double.eps}) to be same as original function}

\item{signlogp_base}{integer; log base when calculating signed log_base p metric}

\item{get_terms}{logical; \code{TRUE} to run \code{Rubrary::get_GSEAsq_terms} to output filtered table of terms and associated statistics}

\item{terms_filt_freq}{num vector length 2; \code{get_GSEAsq_terms} arg, \code{filt_freq[1]} is minimum frequency, \code{filt_freq[2]} is maximum frequency}

\item{plot_type}{\code{c("jitter", "density")}; GSEA squared output plot type}

\item{plot_pval}{logical; \code{TRUE} to append p-value to category name. If \code{ggtext} package is installed, can do fancy markdown formatting}

\item{title}{string; plot title}

\item{cat_colors}{char vector; list of colors corresponding to category}

\item{savename}{string; path to save outputs under (no extension)}

\item{plot_fmt}{string; file extension to save plot as}

\item{height}{numeric; output plot height}

\item{width}{numeric; output plot width}

\item{seed}{integer; randomization seed}
}
\value{
\code{GSEAsq} object with \code{ggplot} object \code{plot}, df of original GSEA input with categories \code{pathways}, df of categories KS statistics \code{categories}, and df of terms KS statistics \code{terms} if applicable
}
\description{
Run GSEA Squared on GSEA results
}
\examples{
library(dplyr)
# Load data
airway_deseq_res <- Rubrary::airway_deseq_res
deseq_stats <- setNames(
  airway_deseq_res[,"sign_log_p"],
  airway_deseq_res[,"hgnc_symbol"]
)
pthwys <- Rubrary::GSEA_pathways
# Run (f)GSEA
gsea_results <- fgsea::fgsea(
  pathways = pthwys,
  stats = deseq_stats,
  eps = 0.0,
  minSize = 15,
  maxSize  = 500) \%>\%
  arrange(NES)
# Run GSEA Squared
GSEAsq_terms <- c("METABOLIC","DNA")
# Run GSEA squared
GSEAsq <- Rubrary::run_GSEA_squared(
  df_GSEA = gsea_results,
  get_terms = TRUE, verbose = FALSE,
  categories = GSEAsq_terms,
  cat_terms = GSEAsq_terms,
  plot_pval = TRUE,
  plot_type = "jitter"
)
names(GSEAsq) # Various outputs as list
GSEAsq$plot
}
