% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lisi.R
\name{run_lisi}
\alias{run_lisi}
\title{Compute and plot local inverse Simpson index}
\usage{
run_lisi(
  sobj,
  labels,
  reduction = c("umap", "pca", "tsne"),
  title = NULL,
  savename = NULL
)
}
\arguments{
\item{sobj}{Seurat object}

\item{labels}{character vector; variable colname in metadata to calculate LISI values against}

\item{reduction}{c("umap", "pca", "tsne"); dimensional reduction method}

\item{title}{string; plot title}

\item{savename}{string; filepath to save results and figure under (no ext.)}
}
\value{
Dataframe of LISI values
}
\description{
Wrapper for \href{https://github.com/immunogenomics/LISI}{immunogenomics/LISI} that takes in Seurat object, calculates LISI per cell for given labels/variables, and plots visualization of LISI results.
}
