% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PCA.R
\name{run_PCA}
\alias{run_PCA}
\title{Run PCA (prcomp wrapper)}
\usage{
run_PCA(
  df,
  savename = "PCA",
  summary = FALSE,
  center = TRUE,
  scale = TRUE,
  tol = 0.05
)
}
\arguments{
\item{df}{numeric dataframe; samples as columns, genes as rows}

\item{savename}{string; directory path and file name but no extension pls}

\item{summary}{logical; output summary info}

\item{center}{logical; indicate whether the variables should be shifted to be zero centered}

\item{scale}{logical; indicate whether the variables should be scaled to have unit variance before the analysis takes place}

\item{tol}{numerical; indicate the magnitude below which components should be omitted}
}
\value{
output text files of PCA scores, loadings, sdev
}
\description{
Adapted from glab.library "PCA_from_file"
}
