% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_RRHO.R
\name{run_RRHO}
\alias{run_RRHO}
\title{Run RRHO analysis}
\usage{
run_RRHO(sig1, sig2)
}
\arguments{
\item{sig1}{string/dataframe;}

\item{sig2}{string/dataframe;}
}
\description{
Inference on the amount of agreement in two sorted lists using the Rank-Rank Hypergeometric Overlap test.
Outputs RRHO results, hypermatrix heatmap, rank rank scatter, metric scatter
}
