% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rwrite}
\alias{rwrite}
\title{Wrapper for \code{data.table::fwrite} with tab separated values (TSV) default parameters}
\usage{
rwrite(x, file, sep = "\\t", quote = FALSE, row.names = FALSE)
}
\arguments{
\item{x}{dataframe; \code{fwrite} input}

\item{file}{string; filepath to save output under, with extension}

\item{sep}{string; separator between columns}

\item{quote}{string; T to wrap fields in quotes}

\item{row.names}{logical; T to write rownames to file}
}
\value{
Dataframe written out to \code{savename} file
}
\description{
Wrapper for \code{data.table::fwrite} with tab separated values (TSV) default parameters
}
