% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_line}
\alias{split_line}
\title{Split text into multiple lines}
\usage{
split_line(text, chars = 40, lines = NULL)
}
\arguments{
\item{text}{string}

\item{chars}{integer; rough # of characters per line}

\item{lines}{integer; desired # of lines}
}
\value{
\code{text} split into lines if longer than \code{split_nchar}
}
\description{
Split text into multiple lines
}
\examples{
lipsum <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."
writeLines(split_line(lipsum, chars = 50))
writeLines(split_line(lipsum, lines = 3))
writeLines(split_line("Short line of text!"))
writeLines(split_line("Short line of text!", lines = 2))
}
