% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_integration.R
\name{assess_integration}
\alias{assess_integration}
\alias{run_LISI}
\alias{run_MixingMetric}
\title{Compute and plot integration assessment methods for Seurat object}
\usage{
assess_integration(
  sobj,
  group.var,
  method = c("LISI", "MixingMetric"),
  reduction = c("umap", "pca", "tsne"),
  MM_flip = TRUE,
  title = NULL,
  savename = NULL,
  width = NULL,
  height = NULL
)

run_LISI(
  sobj,
  group.var,
  reduction = c("umap", "pca", "tsne"),
  title = NULL,
  savename = NULL,
  width = NULL,
  height = NULL
)

run_MixingMetric(
  sobj,
  group.var,
  MM_flip = TRUE,
  reduction = c("umap", "pca", "tsne"),
  title = NULL,
  savename = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{sobj}{Seurat object}

\item{group.var}{character (vector); metadata colname to check level of integration}

\item{method}{c("LISI", "MixingMetric"); integration assessment metric to apply}

\item{reduction}{c("umap", "pca", "tsne"); dimensional reduction method}

\item{MM_flip}{logical; use \code{max.k - MixingMetric} instead of \code{MixingMetric} only}

\item{title}{string; plot title}

\item{savename}{string; filepath to save results and figure under (no ext.)}

\item{width}{numeric; plot width}

\item{height}{numeric; plot height}
}
\value{
Dataframe of integration assessment values
}
\description{
Wrapper for application of single cell integration assessment metrics with default parameters. Takes in Seurat object, calculates integration assessment metric per cell for given grouping variables, and plots visualization of metrics. A single integration score per dataset can be reported as the median of the metric across all cells.
}
\details{
\subsection{LISI: local inverse Simpson's index}{

Wrapper for \href{https://github.com/immunogenomics/LISI}{immunogenomics/LISI} from Harmony paper. Quantifies diversity of batches within local neighborhoods, defined by kNN graphs, using the inverse Simpson's index.
\itemize{
\item LISI = 1: neighborhood comprised of single group, not mixed
\item LISI = # of groups: neighborhood has equal number of cells from all datasets, perfect mixing
}
}

\subsection{Seurat MixingMetric: local / group kNN ranking}{

Wrapper for \code{Seurat::MixingMetric} from Seurat v3 paper. Assesses local neighborhood composition for a cell and how well mixed that neighborhood is. If the neighborhood is well mixed, should contain at least some cells from each dataset. Can be more intuitive to have high MixingMetric mean good mixing, so range can be flipped with \code{max.k - MixingMetric}. Uses flipped scale (\code{MM_flip = TRUE}) by default.
\itemize{
\item MixingMetric = \code{max.k}, low \code{max.k - MixingMetric}: not mixed
\item Low MixingMetric, high \code{max.k - MixingMetric}: well mixed
}
}
}
\section{Functions}{
\itemize{
\item \code{run_LISI()}: Compute and plot local inverse Simpson's index (LISI)

\item \code{run_MixingMetric()}: Compute and plot Seurat MixingMetric

}}
\seealso{
\code{\link[Seurat:MixingMetric]{Seurat::MixingMetric()}}
}
