% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq.R
\name{df_to_mtx}
\alias{df_to_mtx}
\title{df_to_mtx}
\usage{
df_to_mtx(df, merged = T, filter = T)
}
\arguments{
\item{df}{A dataframe}

\item{merged}{If dataframe is made from combined datasets}

\item{filter}{Apply filter for genes with 9+ zeros}
}
\value{
Dataframe restructured as matrix
}
\description{
df_to_mtx
}
