% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter dataframe by list of (protein coding) genes}
\usage{
filter_genes(
  df,
  genes_filt = Rubrary::get_PC_genes(),
  gene_col = NULL,
  search = FALSE
)
}
\arguments{
\item{df}{dataframe; rownames must be gene symbols}

\item{genes_filt}{char vector; (protein coding) genes to filter by}

\item{gene_col}{string; colnames of genes, assumed \code{rownames(df)} if NULL}

\item{search}{logical; T to use \code{Seurat::UpdateSymbolList} to match gene symbols better}
}
\value{
dataframe with gene rownames filtered to protein coding only
}
\description{
Filter df (genes in rownames) to list of genes given.
}
