% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSEA.R
\name{format_GSEA_name}
\alias{format_GSEA_name}
\title{Format GSEA pathway name}
\usage{
format_GSEA_name(
  pw,
  ignore = c("DNA", "RNA", "NADH", " IV "),
  split = FALSE,
  split_nchar = 40
)
}
\arguments{
\item{pw}{string; MSigDB pathway name}

\item{ignore}{char vector; terms to leave as all uppercase (ex. DNA, RNA) instead of title case}

\item{split}{logical; (roughly) split long names into multiple line?}

\item{split_nchar}{integer; if \code{split = TRUE}, max number of characters per line}
}
\value{
Nicer looking title-case'd pathway name
}
\description{
Strips prepended database abbreviation, changes underscores to spaces, and converts to title case.
}
\examples{
format_GSEA_name("DATABASE_DNA_REPAIR")
format_GSEA_name("DATABASE_DNA_REPAIR_THIS_IS_A_REALLY_LONG_PATHWAY", split = TRUE)
}
