% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_PC_genes}
\alias{get_PC_genes}
\title{Retrieve protein coding genes from Ensembl BioMart}
\usage{
get_PC_genes()
}
\value{
character vector of protein coding genes
}
\description{
Retrieve protein coding genes from Ensembl BioMart
}
\examples{
\donttest{
  head(Rubrary::get_PC_genes())
}
}
