% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq_old.R
\name{make_pconly}
\alias{make_pconly}
\title{Filter signature to PC coding only by matching with genes in log2UQcounts}
\usage{
make_pconly(sig_path)
}
\arguments{
\item{sig_path}{string; path to DESeq signature}
}
\value{
DESeq signature subsetted to PC only genes
}
\description{
Filter signature to PC coding only by matching with genes in log2UQcounts
}
