% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_DESeq.R
\name{output_DESeq}
\alias{output_DESeq}
\title{Output DESeq results as text and rank file}
\usage{
output_DESeq(DE_results, outname = "DESeq_results")
}
\arguments{
\item{DE_results}{DESeq results object}

\item{outname}{string; output filename sans extenstion}
}
\value{
DESeq results as text and rank file
}
\description{
Output DESeq results as text and rank file
}
