% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgenes.R
\name{output_DESeq}
\alias{output_DESeq}
\title{Output DESeq results as dataframe with signed log p metric}
\usage{
output_DESeq(DE_results, logbase = 2, savename = NULL, rank = FALSE)
}
\arguments{
\item{DE_results}{\code{DESeqDataSet} object}

\item{logbase}{numeric; base of log for sign log_base p metric}

\item{savename}{string; filepath to save table under}

\item{rank}{logical; \code{TRUE} to output .rnk file compatible with GSEA Java app}
}
\value{
Dataframe with DESeq results including signed log p metric
}
\description{
Filters out all \code{sign_log_p == NA} rows.
}
