% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_barplot.R
\name{plot_comp_barplot}
\alias{plot_comp_barplot}
\title{Plot stacked barplot filled by group}
\usage{
plot_comp_barplot(
  sobj,
  breaks,
  group,
  stack = TRUE,
  counts = TRUE,
  title = ggplot2::waiver(),
  xlabel = breaks,
  colors = NULL,
  ncol_legend = NULL,
  savename = NULL,
  width = 7,
  height = 7
)
}
\arguments{
\item{sobj}{Seurat object}

\item{breaks}{string; metadata colname in sobj to set breaks / columns}

\item{group}{string; metadata colname in sobj to split columns by}

\item{stack}{logical; split into stacked plots of counts + fraction}

\item{counts}{logical; T to include counts barplot above stacked barplot}

\item{title}{string; plot title}

\item{xlabel}{string; x-axis (breaks) label}

\item{colors}{char vector; list of colors (n = # of unique groups), or "alpha"}

\item{ncol_legend}{integer; # of legend columns}

\item{savename}{string; filepath to save plot under}

\item{width}{numeric; saved plot width}

\item{height}{numeric; saved plot height}
}
\value{
Stacked barplot, filled by group composition
}
\description{
Plot stacked barplot filled by group
}
