% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Density plot by group}
\usage{
plot_density(
  df,
  value,
  group,
  group2 = NA,
  title = NA,
  pval = T,
  xlab = value,
  colors = c("firebrick3", "gray"),
  rug = F,
  rug_label = F,
  rug_lab_nudge = 1e-05,
  savename = NA,
  width = 6,
  height = 4
)
}
\arguments{
\item{df}{dataframe}

\item{value}{string; colname of values}

\item{group}{string; colname of group info}

\item{group2}{string; group of interest}

\item{title}{string; plot title}

\item{pval}{logical; get KSpval first group vs others}

\item{xlab}{string; x-axis label}

\item{colors}{vector; # of colors = to # of groups}

\item{rug}{logical; put rug plot on x axis}

\item{rug_label}{logical; label rug ticks with first df column}

\item{rug_lab_nudge}{numeric; fine-tune distance of label from x-axis}

\item{savename}{string; filepath to save PNG under}

\item{width}{numeric; ggsave width of plot}

\item{height}{numeric; ggsave height of plot}
}
\value{
Density plot by group
}
\description{
Density plot by group
}
\examples{
set.seed(13)

}
