% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Density plot by group}
\usage{
plot_density(
  df,
  value,
  group,
  pval = T,
  colors = c("gray", "firebrick3"),
  savename = NA
)
}
\arguments{
\item{df}{dataframe}

\item{value}{string; colname of values}

\item{group}{string; colname of group info}

\item{pval}{logical; get KSpval first group vs others}

\item{colors}{vector; # of colors = to # of groups}

\item{savename}{string; filepath to save PNG under}
}
\value{
Density plot by group
}
\description{
Density plot by group
}
\examples{
set.seed(13)

}
