% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DESeq_boxplot.R
\name{plot_DESeq_boxplot}
\alias{plot_DESeq_boxplot}
\title{Ashvath's DESeq boxplot code in function format}
\usage{
plot_DESeq_boxplot(sig, data, anno, batch)
}
\arguments{
\item{sig}{DESeq signature}

\item{data}{log2UQ transformed data}

\item{anno}{annotation file}

\item{batch}{short descriptor for set that will get appended to filename}
}
\description{
Ashvath's DESeq boxplot code in function format
}
