% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot distribution of numeric vector}
\usage{
plot_distribution(
  values,
  check_normal = FALSE,
  hist = FALSE,
  title = "Distribution",
  xlab = "Value",
  savename = NULL,
  height = 5,
  width = 7
)
}
\arguments{
\item{values}{numeric vector; values to check distribution for}

\item{check_normal}{logical; perform SW test for normality?}

\item{hist}{logical; include histogram?}

\item{title}{string; plot title}

\item{xlab}{string; x-axis label}

\item{savename}{string; filepath to save figure under}

\item{height}{numeric; plot height}

\item{width}{numeric; plot width}
}
\value{
Plot of distribution with median + mean indicated
}
\description{
Ignores NA values for mean & median calculation.
}
\examples{
set.seed(13)
vals_normal <- rnorm(100)
plot_distribution(values = vals_normal, title = "Normal", hist = TRUE, check_normal = TRUE)

vals_sequential <- c(1:100)
plot_distribution(values = vals_sequential, title = "Sequential", check_normal = TRUE)

}
\seealso{
\code{\link[=check_normal]{check_normal()}}
}
