% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSEA_pathway.R
\name{plot_GSEA_pathway}
\alias{plot_GSEA_pathway}
\title{Plot GSEA pathway plots (waterfall + enrichment)}
\usage{
plot_GSEA_pathway(
  sig,
  rankcol,
  rankcol_name = rankcol,
  geneset,
  geneset_name,
  subtitle = NA,
  savename = NA
)
}
\arguments{
\item{sig}{dataframe; "gene" + rankcol columns}

\item{rankcol}{string; colname of values}

\item{rankcol_name}{string; descriptive name of values}

\item{geneset}{vector; list of genes}

\item{geneset_name}{string; name of geneset}

\item{subtitle}{string; (name of signature?)}

\item{savename}{string; filepath to save png under}
}
\value{
grid of gene set enrichment waterfall above enrichment plot
}
\description{
Plot GSEA pathway plots (waterfall + enrichment)
}
