% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{plot_GSEA_pathway}
\alias{plot_GSEA_pathway}
\title{Plot GSEA pathway plots (waterfall + enrichment)}
\usage{
plot_GSEA_pathway(
  sig,
  geneset,
  genecol = "gene",
  rankcol,
  rankcol_name = rankcol,
  hightolow = FALSE,
  label = length(geneset) < 20,
  legendpos = "none",
  hl_color = "firebrick3",
  lab_high = NULL,
  lab_low = NULL,
  hllab = "Highlight",
  title = NULL,
  subtitle = NULL,
  savename = NULL
)
}
\arguments{
\item{sig}{df/string; (path to) dataframe with genecol + rankcol columns}

\item{geneset}{vector; list of genes}

\item{genecol}{string; colname of gene names in \code{sig}}

\item{rankcol}{string; colname of values in \code{sig}}

\item{rankcol_name}{string; descriptive name of values}

\item{hightolow}{logical; T for high values on left, low on right}

\item{label}{logical; T to label highlighted genes}

\item{legendpos}{vector; value btwn 0-1 as legend coordinates (ggplot2 legend.position)}

\item{hl_color}{string; color for highlight}

\item{lab_high}{string; description of high values}

\item{lab_low}{string; description of low values}

\item{hllab}{string; description of highlighted values}

\item{title}{string; plot title}

\item{subtitle}{string; plot subtitle; !! overwrites p-val_enrichment}

\item{savename}{string; filepath to save png under}
}
\value{
grid of gene set enrichment waterfall above enrichment plot
}
\description{
If subtitle is not specified, Kolmogorov-Smirnov enrichment p-value is calculated to check if the genes in the pathway separate from the genes not in the chosen pathway based on ranking by the desired \code{rankcol}.
}
\examples{
airway_deseq = Rubrary::airway_deseq_res
genes = Rubrary::GSEA_pathways$GOBP_REGULATION_OF_GLUCOSE_IMPORT
Rubrary::plot_GSEA_pathway(
  sig = airway_deseq,
  geneset = genes,
  genecol = "hgnc_symbol",
  rankcol = "sign_log_p",
  rankcol_name = "Sign log p value",
  lab_high = "\U2191 in treated\n\U2193 in untreated",
  lab_low = "\U2191 in untreated\n\U2193 in treated",
  hllab = "Highlight genes",
  title = "Airway Treated vs. Untreated DESeq - Glucose Import Regulation Genes"
)

}
