% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSEA_pathway.R
\name{plot_GSEA_pathway}
\alias{plot_GSEA_pathway}
\title{Plot GSEA pathway plots (waterfall + enrichment)}
\usage{
plot_GSEA_pathway(
  sig,
  geneset,
  rankcol,
  rankcol_name = rankcol,
  hightolow = FALSE,
  label = length(geneset) < 20,
  legendpos = "none",
  highlab = NA,
  lowlab = NA,
  hllab = "Highlight",
  title = "",
  subtitle = NULL,
  savename = NULL
)
}
\arguments{
\item{sig}{dataframe; "gene" + rankcol columns}

\item{geneset}{vector; list of genes}

\item{rankcol}{string; colname of values}

\item{rankcol_name}{string; descriptive name of values}

\item{hightolow}{logical; T for high values on left, low on right}

\item{label}{logical; T to label highlighted genes}

\item{legendpos}{vector; value btwn 0-1 as legend coordinates (ggplot legend.position option)}

\item{highlab}{string; description of high values}

\item{lowlab}{string; description of low values}

\item{hllab}{string; description of highlighted values}

\item{title}{string; plot title}

\item{subtitle}{string; plot subtitle; !! overwrites p-val_enrichment}

\item{savename}{string; filepath to save png under}
}
\value{
grid of gene set enrichment waterfall above enrichment plot
}
\description{
Plot GSEA pathway plots (waterfall + enrichment)
}
