% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSEA_pathway.R
\name{plot_GSEA_pathway}
\alias{plot_GSEA_pathway}
\title{Plot GSEA pathway plots (waterfall + enrichment)}
\usage{
plot_GSEA_pathway(
  sig,
  geneset,
  rankcol,
  rankcol_name = rankcol,
  highlab = NA,
  lowlab = NA,
  hllab = "Highlight",
  title = "",
  subtitle = NA,
  savename = NA
)
}
\arguments{
\item{sig}{dataframe; "gene" + rankcol columns}

\item{geneset}{vector; list of genes}

\item{rankcol}{string; colname of values}

\item{rankcol_name}{string; descriptive name of values}

\item{highlab}{string; description of high values}

\item{lowlab}{string; description of low values}

\item{hllab}{string; description of highlighted values}

\item{title}{string; title of plot}

\item{subtitle}{string; (name of signature?)}

\item{savename}{string; filepath to save png under}
}
\value{
grid of gene set enrichment waterfall above enrichment plot
}
\description{
Plot GSEA pathway plots (waterfall + enrichment)
}
