% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSEA.R
\name{plot_GSEA_pathway_batch}
\alias{plot_GSEA_pathway_batch}
\title{\code{plot_GSEA_pathway} that works nicely with \code{lapply}}
\usage{
plot_GSEA_pathway_batch(
  path_name,
  pthwys,
  sig,
  genecol = "gene",
  rankcol,
  rankcol_name = rankcol,
  hllab = "Pathway genes",
  hightolow = FALSE,
  lab_low = NULL,
  lab_high = NULL,
  legendpos = c(0.5, 0.2),
  label = length(pthwys[[path_name]]) < 50,
  sig_name = "",
  savedir = NULL
)
}
\arguments{
\item{path_name}{string; name of pathway}

\item{pthwys}{named list; key = geneset name, values = char vector of genes in geneset}

\item{sig}{dataframe; signature}

\item{genecol}{string; colname of gene names in \code{sig}}

\item{rankcol}{string; colname of values to rank by}

\item{rankcol_name}{string; descriptor of rankcol}

\item{hllab}{string; descriptor of highlighted genes}

\item{hightolow}{logical; T for high values on left, low on right}

\item{lab_low}{string; label for low rankcol values}

\item{lab_high}{string; label for high rankcol values}

\item{legendpos}{vector; value btwn 0-1 as legend coordinates (ggplot2 legend.position)}

\item{label}{logical; T to label points in plot}

\item{sig_name}{string; name of signature}

\item{savedir}{string; directory path for saving plot}
}
\value{
Gene set enrichment plot as ggplot object
}
\description{
\code{plot_GSEA_pathway} that works nicely with \code{lapply}
}
