% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAsq.R
\name{plot_GSEAsq_density}
\alias{plot_GSEAsq_density}
\title{Plot categorical percentile rank comparison between two GSEA squared signatures}
\usage{
plot_GSEAsq_density(
  GSEAsq_df1,
  GSEAsq_df2,
  name1,
  name2,
  plot_pval = TRUE,
  cat_order = NULL,
  title = "GSEA Squared",
  colors = c("firebrick", "lightslateblue"),
  rug = TRUE,
  savename = NULL,
  plot_fmt = "png",
  height = 8,
  width = 8
)
}
\arguments{
\item{GSEAsq_df1}{df/string; "gsea_squared-df.txt" df or path for GSEAsq result 1}

\item{GSEAsq_df2}{df/string; "gsea_squared-df.txt" df or path for GSEAsq result 2}

\item{name1}{string; descriptor for GSEAsq result 1}

\item{name2}{string; descriptor for GSEAsq result 2}

\item{plot_pval}{logical; \code{TRUE} to include KS p-value per category in plot}

\item{cat_order}{vector; desired order of GSEAsq categories}

\item{title}{string; overall plot title}

\item{colors}{vector; first color to signify result 1, second for result 2}

\item{rug}{logical; TRUE for rugplot below density plot}

\item{savename}{string; filepath to save outputs under (no extension)}

\item{plot_fmt}{string; file extension to save plot as}

\item{height}{numeric; plot height}

\item{width}{numeric; plot width}
}
\value{
facet wrapped ggplot object
}
\description{
Plot categorical percentile rank comparison between two GSEA squared signatures
}
