% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA with annotation}
\usage{
plot_PCA(
  df,
  dfanno = NA,
  PCx = "PC1",
  PCy = "PC2",
  PCtype = "Loading",
  label = T,
  annoname = "Sample",
  annotype = "Batch",
  title = paste0("PCA ", PCtype, " Plot - ", annotype),
  marginal = F,
  save = F,
  savename = "PCA_plot.png"
)
}
\arguments{
\item{df}{DF or path to PC output}

\item{dfanno}{Annotation info for DF}

\item{PCx}{Component on x-axis}

\item{PCy}{Component on y-axis}

\item{PCtype}{"Loading" or "Scores"}

\item{label}{T to label points}

\item{annoname}{Colname in dfanno matching point name}

\item{annotype}{Colname in dfanno with annotation}

\item{title}{Plot title}

\item{marginal}{Marginal density plot along both axes}

\item{save}{T to save to png}

\item{savename}{Output plot name}
}
\value{
PCA output plotted with annotation
}
\description{
Plot PCA with annotation
}
