% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA with annotation}
\usage{
plot_PCA(
  df_pca,
  anno = NULL,
  PCx = "PC1",
  PCy = "PC2",
  type = c("Scores", "Loadings"),
  label = FALSE,
  annoname = "Sample",
  annotype = "Batch",
  annotype2 = NULL,
  highlight = NULL,
  colors = NULL,
  title = NULL,
  subtitle = NULL,
  density = FALSE,
  savename = NULL,
  width = 8,
  height = 8
)
}
\arguments{
\item{df_pca}{string or prcomp obj; (path to) PCA output}

\item{anno}{string or df; Annotation info for DF}

\item{PCx}{string; Component on x-axis}

\item{PCy}{string; Component on y-axis}

\item{type}{c("Score", "Loading")}

\item{label}{logical; T to label points}

\item{annoname}{string; Colname in \code{anno} matching point name}

\item{annotype}{string; Colname in \code{anno} with info to color by}

\item{annotype2}{string; Colname in \code{anno} with info to change shape by}

\item{highlight}{char vector; Specific points to shape differently & label}

\item{colors}{char vector; length should be number of unique \code{annotype}s}

\item{title}{string; Plot title}

\item{subtitle}{string; Subtitle for plot}

\item{density}{logical; Show density plot along both axes}

\item{savename}{string; filepath to save plot under}

\item{width}{numeric; plot width}

\item{height}{numeric; plot width}
}
\value{
PCA output plotted with annotation
}
\description{
Plot PCA with annotation
}
\examples{
data(iris)
iris$Sample = rownames(iris)
PCA_iris <- Rubrary::run_PCA(t(iris[,c(1:4)]))
# Scores
Rubrary::plot_PCA(df_pca = PCA_iris, anno = iris[,c("Sample", "Species")], annoname = "Sample", annotype = "Species", title = "Iris PCA Scores by Species")
# Loadings
Rubrary::plot_PCA(df_pca = PCA_iris, type = "Loadings", title = "Iris PCA Loadings", label = T)

}
