% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot simple scatter}
\usage{
plot_scatter(
  df = NA,
  xval,
  yval,
  eqlims = F,
  save = F,
  outname = "Scatter.png"
)
}
\arguments{
\item{df}{Dataframe; includes both signatures}

\item{xval}{String; colname for x axis}

\item{yval}{String; colname for y axis}

\item{eqlims}{Equal limits? TO BE IMPLEMENTED}

\item{save}{Logical; save as .png}

\item{outname}{Output file name}
}
\value{
Simple scatter plot
}
\description{
Plot simple scatter
}
