% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot simple scatter}
\usage{
plot_scatter(df, xval, yval, eqlims = F)
}
\arguments{
\item{df}{Dataframe}

\item{xval}{Colname for x axis}

\item{yval}{Colname for y axis}

\item{eqlims}{Equal limits?}
}
\value{
Simple scatter plot
}
\description{
Plot simple scatter
}
