% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot simple scatter}
\usage{
plot_scatter(
  df = NA,
  xval,
  yval,
  label = NA,
  eqlims = F,
  title = paste0(xval, " vs. ", yval),
  outname = NA
)
}
\arguments{
\item{df}{Dataframe; includes both signatures}

\item{xval}{String; colname for x axis}

\item{yval}{String; colname for y axis}

\item{label}{String; colname for point labels}

\item{eqlims}{Equal limits? TO BE IMPLEMENTED}

\item{title}{String; title of plot}

\item{outname}{Output file name}
}
\value{
Simple scatter plot
}
\description{
Plot simple scatter
}
\examples{
df <- data.frame(A = 1:10, B = 11:20, C = LETTERS[1:10])
plot_scatter(df, xval = "A", yval = "B", label = "C", title = "Example Scatter")

}
