% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot simple scatter}
\usage{
plot_scatter(
  df = NULL,
  xval,
  yval,
  label = NULL,
  group = NULL,
  rank = FALSE,
  xlabel = ifelse(methods::is(xval, "character"), xval, "X"),
  ylabel = ifelse(methods::is(yval, "character"), yval, "Y"),
  cormethod = c("pearson", "spearman", "none"),
  guides = TRUE,
  pt_size = 2,
  pt_alpha = 1,
  lbl_size = 3,
  heatmap = FALSE,
  hm_palette = "Spectral",
  reverse = FALSE,
  title = NULL,
  subtitle = NULL,
  savename = NULL,
  width = 8,
  height = 8
)
}
\arguments{
\item{df}{dataframe; includes both signatures}

\item{xval}{string/numeric vector; colname for x axis}

\item{yval}{string/numeric vector; colname for y axis}

\item{label}{string; colname for point labels}

\item{group}{string; colname for group color}

\item{rank}{logical; change metric to rank}

\item{xlabel}{string; xval description}

\item{ylabel}{string; yval description}

\item{cormethod}{string; correlation method for stats::cor() or "none"}

\item{guides}{logical; T to include fitted linear model line}

\item{pt_size}{numeric; point size}

\item{pt_alpha}{numeric; point alpha value}

\item{lbl_size}{numeric; label text size}

\item{heatmap}{logical; T for underlying heatmap (untested)}

\item{hm_palette}{string; RColorBrewer continuous palette name (untested)}

\item{reverse}{logical; T to reverse both axes}

\item{title}{string; plot title}

\item{subtitle}{string; plot subtitle}

\item{savename}{string; filepath to save figure under}

\item{width}{numeric; plot width}

\item{height}{numeric; plot height}
}
\value{
Simple scatter plot as ggplot2
}
\description{
Correlation value and method gets placed in plot caption / bottom-right.
}
\examples{
df <- data.frame(A = 1:10, B = 11:20, C = LETTERS[1:10], D = rep(LETTERS[1:5], each = 2))
plot_scatter(df, xval = "A", yval = "B", label = "C", group = "D", title = "Example Scatter")

}
