% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot simple scatter}
\usage{
plot_scatter(
  df = NULL,
  xval,
  yval,
  label = NULL,
  rank = FALSE,
  xlabel = ifelse(methods::is(xval, "character"), xval, "X"),
  ylabel = ifelse(methods::is(yval, "character"), xval, "Y"),
  cormethod = c("pearson", "spearman"),
  guides = TRUE,
  alpha = 1,
  heatmap = FALSE,
  hm_palette = "Spectral",
  reverse = FALSE,
  title = NULL,
  subtitle = NULL,
  savename = NULL
)
}
\arguments{
\item{df}{dataframe; includes both signatures}

\item{xval}{string/numeric vector; colname for x axis}

\item{yval}{string/numeric vector; colname for y axis}

\item{label}{string; colname for point labels}

\item{rank}{logical; change metric to rank}

\item{xlabel}{string; xval description}

\item{ylabel}{string; yval description}

\item{cormethod}{string; correlation method for stats::cor()}

\item{guides}{logical; T to include fitted line}

\item{alpha}{numeric; point alpha value}

\item{heatmap}{logical; T for underlying heatmap (untested)}

\item{hm_palette}{string; RColorBrewer continuous palette name (untested)}

\item{reverse}{logical; T to reverse both axes}

\item{title}{string; plot title}

\item{subtitle}{string; plot subtitle}

\item{savename}{string; filepath to save figure under}
}
\value{
Simple scatter plot as ggplot2
}
\description{
Correlation value and method gets placed in plot caption / bottom-right.
}
\examples{
df <- data.frame(A = 1:10, B = 11:20, C = LETTERS[1:10])
plot_scatter(df, xval = "A", yval = "B", label = "C", title = "Example Scatter")

}
