% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_compare.R
\name{plot_scatter_compare}
\alias{plot_scatter_compare}
\title{Plot scatter, either metric or rank}
\usage{
plot_scatter_compare(
  set1path,
  set2path,
  set1lab = "Set 1",
  set2lab = "Set 2",
  rank = F,
  guides = F,
  title = paste0(set1lab, " vs. ", set2lab),
  subtitle = "",
  savename = NULL
)
}
\arguments{
\item{set1path}{Path to 1st set (x axis)}

\item{set2path}{Path to 2nd set (y axis)}

\item{set1lab}{Axis label for 1st set (x)}

\item{set2lab}{Axis label for 2nd set (y)}

\item{rank}{Convert sign log p values to rank}

\item{guides}{logical; Show m=1 and linear fit?}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{savename}{string; File output full name}
}
\value{
Plot comparing DESeq signatures
}
\description{
Rubrary::plot_scatter is written to be a bit more general.
}
