% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SCN_KaplanMeier.R
\name{plot_SCN_KaplanMeier}
\alias{plot_SCN_KaplanMeier}
\title{Plot SCN Kaplan Meier Survival Analysis}
\usage{
plot_SCN_KaplanMeier(
  df,
  xlab = "Time",
  title = "Survival Analysis",
  SCN = NULL,
  savename = NULL
)
}
\arguments{
\item{df}{dataframe; cols "OS" (overall survival censor), "OS.time" (overall survival time)}

\item{xlab}{string; x-axis label}

\item{title}{plot title}

\item{SCN}{SCN threshold to designate SCN-like status if not already in DF}

\item{savename}{string; filename to save as}
}
\value{
Kaplan-Meier overall survival analysis plot comparing SCN-like status
}
\description{
NOT GENERALIZED
}
