% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PCA.R
\name{plot_screeplot}
\alias{plot_screeplot}
\title{Plot screeplot from prcomp}
\usage{
plot_screeplot(obj_prcomp, savename = NA, label = F)
}
\arguments{
\item{obj_prcomp}{prcomp function output object}

\item{savename}{string; filepath to save PNG under}

\item{label}{logical; T for \% values at points}
}
\value{
Screeplot
}
\description{
Plot screeplot from prcomp
}
