% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot waterfall plot (with optional density)}
\usage{
plot_waterfall(
  sig,
  highlight,
  rankcol,
  rankcol_name = rankcol,
  label = TRUE,
  vert = FALSE,
  density = FALSE,
  hllab = "Highlight",
  otherlab = "Others",
  pval = TRUE,
  highlab = NA,
  lowlab = NA,
  legendpos = "none",
  title = NULL,
  colors = c("firebrick3", "gray"),
  width = 10,
  height = 5,
  savename = NULL
)
}
\arguments{
\item{sig}{dataframe; name (1st column) + rankcol columns}

\item{highlight}{vector; list of names to highlight in waterfall}

\item{rankcol}{string; colname of values}

\item{rankcol_name}{string; descriptive name of values}

\item{label}{logical; T to label highlighted values}

\item{vert}{logical; T for columns to be horizontal (not optimized)}

\item{density}{logical; T to output aligned density plot w/ wf}

\item{hllab}{string; description of highlighted values}

\item{otherlab}{string; description of non-highlighted values}

\item{pval}{logical; T to include KS enrichment pvalue}

\item{highlab}{string; descriptor for high rankcol values}

\item{lowlab}{string; descriptor for low rankcol values}

\item{legendpos}{vector; value btwn 0-1 as legend coordinates (ggplot legend.position option)}

\item{title}{string; plot title}

\item{colors}{vector; two colors, for highlight vs. other}

\item{width}{numeric; width of plot}

\item{height}{numeric; height of plot}

\item{savename}{string; filepath to save figure under}
}
\value{
Waterfall plot
}
\description{
Plot waterfall plot (with optional density)
}
