% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot horizontal waterfall plot (with optional density)}
\usage{
plot_waterfall(
  sig,
  highlight,
  rankcol,
  rankcol_name = rankcol,
  hightolow = FALSE,
  vert = FALSE,
  label = TRUE,
  replab_size = 3.5,
  density = FALSE,
  hllab = "Highlight",
  otherlab = "Others",
  pval = TRUE,
  lab_high = NULL,
  lab_low = NULL,
  lab_size = 6,
  legendpos = "none",
  title = NULL,
  colors = c("firebrick3", "gray"),
  width = 10,
  height = 5,
  savename = NULL
)
}
\arguments{
\item{sig}{dataframe; name (1st column) + rankcol columns}

\item{highlight}{vector; list of names to highlight in waterfall}

\item{rankcol}{string; colname of values}

\item{rankcol_name}{string; descriptive name of values}

\item{hightolow}{logical; T for x-axis by decreasing value}

\item{vert}{(NOT TESTED) logical; T for columns to be horizontal}

\item{label}{logical; T to label highlighted values}

\item{replab_size}{numeric; size of gene label}

\item{density}{logical; T to output aligned density plot w/ wf}

\item{hllab}{string; description of highlighted values}

\item{otherlab}{string; description of non-highlighted values}

\item{pval}{logical; T to include KS enrichment pvalue}

\item{lab_high}{string; descriptor for high rankcol values}

\item{lab_low}{string; descriptor for low rankcol values}

\item{lab_size}{numeric; size of value annotation label}

\item{legendpos}{vector; value btwn 0-1 as legend coordinates (ggplot legend.position option)}

\item{title}{string; plot title}

\item{colors}{vector; two colors, for highlight vs. other}

\item{width}{numeric; width of plot}

\item{height}{numeric; height of plot}

\item{savename}{string; filepath to save figure under}
}
\value{
Waterfall plot
}
\description{
Plot horizontal waterfall plot (with optional density)
}
