% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot waterfall plot (with optional density)}
\usage{
plot_waterfall(
  sig,
  highlight,
  rankcol,
  vert = FALSE,
  density = FALSE,
  ylab = rankcol,
  hllab = "Top SCN",
  otherlab = "Others",
  pval = TRUE,
  highlab = NA,
  lowlab = NA,
  title = NA,
  colors = c("gray", "firebrick3"),
  width = 12,
  height = 6,
  savename = NA
)
}
\arguments{
\item{sig}{dataframe}

\item{highlight}{vector}

\item{rankcol}{string}

\item{vert}{logical}

\item{density}{logical}

\item{ylab}{string}

\item{hllab}{string}

\item{otherlab}{string}

\item{pval}{logical}

\item{highlab}{string}

\item{lowlab}{string}

\item{title}{string}

\item{colors}{vector}

\item{width}{numeric}

\item{height}{numeric}

\item{savename}{string}
}
\value{
Waterfall plot
}
\description{
Plot waterfall plot (with optional density)
}
