% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PCA.R
\name{run_PCA}
\alias{run_PCA}
\title{Run PCA (prcomp wrapper)}
\usage{
run_PCA(
  df,
  savename = NULL,
  summary = FALSE,
  center = TRUE,
  scale = TRUE,
  tol = 0.05,
  screeplot = TRUE
)
}
\arguments{
\item{df}{(path to) numeric dataframe; samples as columns, genes as rows}

\item{savename}{string; filepath (no ext.) to save PCA scores, loadings, sdev under}

\item{summary}{logical; output summary info}

\item{center}{logical; indicate whether the variables should be shifted to be zero centered}

\item{scale}{logical; indicate whether the variables should be scaled to have unit variance before the analysis takes place}

\item{tol}{numerical; indicate the magnitude below which components should be omitted}

\item{screeplot}{logical; output + save screeplot?}
}
\value{
prcomp obj
}
\description{
Adapted from glab.library "PCA_from_file"
}
