% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_RRHO.R
\encoding{UTF-8}
\name{run_RRHO}
\alias{run_RRHO}
\title{Run RRHO analysis}
\usage{
run_RRHO(
  sig1,
  sig2,
  sig1_name,
  sig2_name,
  sig1_low = NA,
  sig1_high = NA,
  sig2_low = NA,
  sig2_high = NA,
  key = "gene",
  metric1 = "sign_log_p",
  metric2 = metric1,
  steps = NULL,
  savename = NULL,
  webtool = TRUE,
  BY = FALSE,
  hm_method = c("ggplot", "lattice"),
  palette = "Spectral",
  waterfall = TRUE,
  scatter = TRUE
)
}
\arguments{
\item{sig1}{string/dataframe; path or df for sig1, with cols \code{key} and \code{metric1}}

\item{sig2}{string/dataframe; path or df for sig2, with cols \code{key} and \code{metric2}}

\item{sig1_name}{string; description of sig1}

\item{sig2_name}{string; description of sig2}

\item{sig1_low}{string; description of low sig1 values}

\item{sig1_high}{string; description of high sig1 values}

\item{sig2_low}{string; description of low sig2 values}

\item{sig2_high}{string; description of high sig2 values}

\item{key}{string; colname of corresponding values btwn both sigs}

\item{metric1}{string; colname of sig1 metric to rank by}

\item{metric2}{string; colname of sig2 metric to rank by}

\item{steps}{integer; RRHO step size}

\item{savename}{string; filepath to save results under (no ext.)}

\item{webtool}{logical; T to output text formatted for Graeber RRHO webtool}

\item{BY}{logical; T for Benjamini-Yekutieli FDR corrected pvalues}

\item{hm_method}{string; make heatmap with ggplot geom_raster or lattice levelplot}

\item{palette}{string; RColorBrewer continuous palette name}

\item{waterfall}{logical; T to include waterfall subplots (plot_method = "ggplot" only)}

\item{scatter}{logical; T to output metric & rank scatterplot}
}
\value{
RRHO results
}
\description{
Inference on the amount of agreement in two sorted lists using the Rank-Rank Hypergeometric Overlap test.
Outputs RRHO results, hypermatrix heatmap, rank rank scatter, metric scatter.
}
\details{
Arrows in unicode are \verb{\\u2193} (down) and \verb{\\u2191} (up) if you want to be cool in your low/high descriptions B)
}
