% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_pkg}
\alias{use_pkg}
\title{Check if package is installed, prompt if not}
\usage{
use_pkg(..., strict = FALSE)
}
\arguments{
\item{...}{string; package(s) to check if installed}

\item{strict}{logical; strict requirements on package install?}
}
\value{
None
}
\description{
Based on https://stackoverflow.com/a/44660688
}
\examples{
use_pkg("ggplot2", "dplyr")
}
